requirements.txt
numpy==1.21.5
Pillow==10.4.0
PyAutoGUI==0.9.53
